+++ Commands/sparkle_help/run_solvers_core.py
@@ -104,8 +104,10 @@ if __name__ == "__main__":
 
     if performance_measure == PerformanceMeasure.QUALITY_ABSOLUTE:
         obj_str = str(quality[0])  # TODO: Handle the multi-objective case
-    else:
+    elif performance_measure == PerformanceMeasure.RUNTIME:
         obj_str = str(cpu_time_penalised)
+    else:
+        print(f"*** ERROR: Unknown performance measure detected: {performance_measure}")
 
     fout = Path(processed_result_path).open("w+")
     fcntl.flock(fout.fileno(), fcntl.LOCK_EX)