+++ Commands/sparkle_help/sparkle_compute_marginal_contribution_help.py
@@ -8,6 +8,7 @@ import os
 import sys
 import csv
 from pathlib import Path
+from typing import Callable
 
 from Commands.sparkle_help import sparkle_basic_help
 from Commands.sparkle_help import sparkle_file_help as sfh
@@ -61,29 +62,18 @@ def write_marginal_contribution_csv(path: Path,
                       "Marginal contributions to the portfolio selector per solver.")
 
 
-def get_capvalue_list(performance_data_csv: SparklePerformanceDataCSV,
-                      performance_measure: PerformanceMeasure) -> list[float] | None:
-    """Return a list of cap-values if the performance measure is QUALITY, else None.
-
-    Args:
-        performance_data_csv: The CSV data as a SparklePerformanceDataCSV object.
-
-    Returns:
-        A list of floating point numbers or None.
-    """
-    # If QUALITY_ABSOLUTE is the performance measure, use the maximum performance per
-    # instance as capvalue; otherwise the cutoff time is used
-    if performance_measure == PerformanceMeasure.QUALITY_ABSOLUTE:
-        return performance_data_csv.get_maximum_performance_per_instance()
-    return None
-
-
 def compute_perfect_selector_marginal_contribution(
+        aggregation_function: Callable[[list[float]], float],
+        capvalue_list: list[float],
+        minimise: bool,
         performance_data_csv_path: Path = sgh.performance_data_csv_path,
         flag_recompute: bool = False) -> list[tuple[str, float]]:
     """Return the marginal contributions of solvers for the VBS.
 
     Args:
+      aggregation_function: function to aggregate the per instance scores
+      capvalue_list: list of cap values
+      minimise: flag indicating if scores should be minimised or maximised
       performance_data_csv_path: Path to the CSV file containing the performance data.
       flag_recompute: Boolean indicating whether a recomputation of the marginal
         contribution is enforced.
@@ -99,21 +89,17 @@ def compute_perfect_selector_marginal_contribution(
               "from file instead! Use --recompute to force recomputation.")
         return read_marginal_contribution_csv(perfect_margi_cont_path)
 
-    cutoff_time_str = str(sgh.settings.get_general_target_cutoff_time())
-    print(f"In this calculation, cutoff time for each run is {cutoff_time_str} seconds")
+    print("In this calculation, cutoff time for each run is "
+          f"{sgh.settings.get_general_target_cutoff_time()} seconds")
 
     rank_list = []
     performance_data_csv = spdcsv.SparklePerformanceDataCSV(performance_data_csv_path)
-    num_instances = performance_data_csv.get_row_size()
-    num_solvers = performance_data_csv.get_column_size()
-    performance_measure = sgh.settings.get_general_performance_measure()
-    capvalue_list = get_capvalue_list(performance_data_csv, performance_measure)
 
     print("Computing virtual best performance for portfolio selector with all solvers "
           "...")
     virtual_best_performance = (
         performance_data_csv.calc_virtual_best_performance_of_portfolio(
-            num_instances, num_solvers, capvalue_list))
+            aggregation_function, minimise, capvalue_list))
     print("Virtual best performance for portfolio selector with all solvers is "
           f"{str(virtual_best_performance)}")
     print("Computing done!")
@@ -124,14 +110,19 @@ def compute_perfect_selector_marginal_contribution(
         tmp_performance_data_csv = spdcsv.SparklePerformanceDataCSV(
             performance_data_csv_path)
         tmp_performance_data_csv.delete_column(solver)
-        tmp_virtual_best_performance = (
+        tmp_virt_best_perf = (
             tmp_performance_data_csv.calc_virtual_best_performance_of_portfolio(
-                num_instances, num_solvers, capvalue_list))
+                aggregation_function, minimise, capvalue_list))
         print("Virtual best performance for portfolio selector excluding solver "
               f"{sfh.get_last_level_directory_name(solver)} is "
-              f"{str(tmp_virtual_best_performance)}")
+              f"{str(tmp_virt_best_perf)}")
         print("Computing done!")
-        marginal_contribution = virtual_best_performance - tmp_virtual_best_performance
+        if minimise and tmp_virt_best_perf > virtual_best_performance or\
+           not minimise and tmp_virt_best_perf < virtual_best_performance:
+            marginal_contribution = tmp_virt_best_perf / virtual_best_performance
+        else:
+            marginal_contribution = 0.0
+
         solver_tuple = (solver, marginal_contribution)
         rank_list.append(solver_tuple)
         print("Marginal contribution (to Perfect Selector) for solver "
@@ -208,8 +199,8 @@ def compute_actual_selector_performance(
         actual_portfolio_selector_path: str,
         performance_data_csv_path: str,
         feature_data_csv_path: str,
-        num_instances: int,
-        num_solvers: int,
+        minimise: bool,
+        aggregation_function: Callable[[list[float]], float],
         capvalue_list: list[float] | None = None) -> float:
     """Return the performance of the selector over all instances.
 
@@ -217,53 +208,42 @@ def compute_actual_selector_performance(
       actual_portfolio_selector_path: Path to portfolio selector.
       performance_data_csv_path: Path to the CSV file with the performance data.
       feature_data_csv_path: path to the CSV file with the features.
-      num_instances: The number of instances.
-      num_solvers: The number of solvers in the portfolio.
+      minimise: Flag indicating, if scores should be minimised.
+      aggregation_function: function to aggregate the performance per instance
       capvalue_list: Optional list of cap-values.
 
     Returns:
       The selector performance as a single floating point number.
     """
-    cutoff_time = sgh.settings.get_general_target_cutoff_time()
     performance_data_csv = spdcsv.SparklePerformanceDataCSV(performance_data_csv_path)
+    penalty_factor = sgh.settings.get_general_penalty_multiplier()
+    performances = []
+    perf_measure = sgh.settings.get_general_performance_measure()
+    capvalue = None
+    for index, instance in enumerate(performance_data_csv.list_rows()):
+        if capvalue_list is not None:
+            capvalue = capvalue_list[index]
+        performance_instance, flag_success = compute_actual_performance_for_instance(
+            actual_portfolio_selector_path, instance, feature_data_csv_path,
+            performance_data_csv, minimise, perf_measure, capvalue)
 
-    actual_selector_performance = 0
+        if not flag_success and capvalue is not None:
+            performance_instance = capvalue * penalty_factor
 
-    for instance_idx in range(0, len(performance_data_csv.list_rows())):
-        instance = performance_data_csv.get_row_name(instance_idx)
+        performances.append(performance_instance)
 
-        if capvalue_list is None:
-            # RUNTIME
-            capvalue = cutoff_time
-            performance_this_instance, flag_successfully_solving = (
-                compute_actual_used_time_for_instance(
-                    actual_portfolio_selector_path, instance, feature_data_csv_path,
-                    performance_data_csv))
-        else:
-            # QUALITY_ABSOLUTE
-            capvalue = capvalue_list[instance_idx]
-            performance_this_instance, flag_successfully_solving = (
-                compute_actual_performance_for_instance(
-                    actual_portfolio_selector_path, instance, feature_data_csv_path,
-                    performance_data_csv))
-
-        if flag_successfully_solving:
-            score_this_instance = (1 + (capvalue - performance_this_instance)
-                                   / (num_instances * num_solvers * capvalue + 1))
-        else:
-            score_this_instance = 0
-
-        actual_selector_performance = actual_selector_performance + score_this_instance
-
-    return actual_selector_performance
+    return aggregation_function(performances)
 
 
 def compute_actual_performance_for_instance(
         actual_portfolio_selector_path: str,
         instance: str,
         feature_data_csv_path: str,
-        performance_data_csv: SparklePerformanceDataCSV) -> tuple[float, bool]:
-    """Return the total time of the selector on a given instance.
+        performance_data_csv: SparklePerformanceDataCSV,
+        minimise: bool,
+        objective_type: PerformanceMeasure,
+        capvalue: float) -> tuple[float, bool]:
+    """Return the actual performance of the selector on a given instance.
 
     Args:
       actual_portfolio_selector_path: Path to the portfolio selector.
@@ -271,87 +251,78 @@ def compute_actual_performance_for_instance(
       feature_data_csv_path: Path to the CSV file with the feature data.
       performance_data_csv: SparklePerformanceDataCSV object that holds the
         performance data.
+      minimise: Whether the performance value should be minimized or maximized
+      objective_type: Whether we are dealing with run time or not.
+      capvalue: Cap value for this instance
 
     Returns:
-      A 2-tuple where the first entry is the numeric performance value and
+      A 2-tuple where the first entry is the performance measure and
       the second entry is a Boolean indicating whether the instance was solved
-      to optimality within the cutoff time.
+      within the cutoff time (Runtime) or at least one solver performance
+      did not exceed capvalue
     """
     feature_data_csv = sfdcsv.SparkleFeatureDataCSV(feature_data_csv_path)
+    # Get the prediction of the selector over the solvers
     list_predict_schedule = get_list_predict_schedule(actual_portfolio_selector_path,
                                                       feature_data_csv, instance)
-    performance_this_instance = sgh.sparkle_maximum_int
-    flag_successfully_solving = True
 
-    for i in range(0, len(list_predict_schedule)):
-        solver = list_predict_schedule[i][0]
-        performance = performance_data_csv.get_value(instance, solver)
-
-        # Take best performance from the scheduled solvers
-        if performance < performance_this_instance:
-            performance_this_instance = performance
-
-    return performance_this_instance, flag_successfully_solving
-
-
-def compute_actual_used_time_for_instance(
-        actual_portfolio_selector_path: str,
-        instance: str,
-        feature_data_csv_path: str,
-        performance_data_csv: SparklePerformanceDataCSV) -> tuple[float, bool]:
-    """Return the best performance of the solver schedule on a given instance.
-
-    Args:
-      actual_portfolio_selector_path: Path to the portfolio selector.
-      instance: Instance name.
-      feature_data_csv_path: Path to the CSV file with the feature data.
-      performance_data_csv: SparklePerformanceDataCSV object that holds the
-        performance data.
-
-    Returns:
-      A 2-tuple where the first entry is the time used  value and
-      the second entry is a Boolean indicating whether the instance was solved
-      to optimality within the cutoff time.
-    """
-    feature_data_csv = sfdcsv.SparkleFeatureDataCSV(feature_data_csv_path)
-    list_predict_schedule = get_list_predict_schedule(actual_portfolio_selector_path,
-                                                      feature_data_csv, instance)
-    cutoff_time = sgh.settings.get_general_target_cutoff_time()
-    used_time_for_this_instance = 0
+    performance = None
     flag_successfully_solving = False
+    cutoff_time = sgh.settings.get_general_target_cutoff_time()
+    if objective_type == PerformanceMeasure.RUNTIME:
+        performance_list = []
+        # In case of Runtime, we loop through the selected solvers
+        for prediction in list_predict_schedule:
+            # A prediction is a solver and its score given by the Selector
+            performance = float(performance_data_csv.get_value(instance, prediction[0]))
+            performance_list.append(performance)
+            scheduled_cutoff_time_this_run = prediction[1]
+            # 1. if performance <= predicted runtime we have a successfull solver
+            if performance <= scheduled_cutoff_time_this_run:
+                # 2. Succes if the tried solvers < selector cutoff_time
+                if sum(performance_list) <= cutoff_time:
+                    flag_successfully_solving = True
+                break
+            # 3. Else, we sett the failed solver to the cutoff time
+            performance_list[-1] = scheduled_cutoff_time_this_run
 
-    for i in range(0, len(list_predict_schedule)):
-        if used_time_for_this_instance >= cutoff_time:
-            flag_successfully_solving = False
-            break
+            # 4. If we have exceeded cutoff_time, we are done
+            if sum(performance_list) > cutoff_time:
+                break
 
-        solver = list_predict_schedule[i][0]
-        scheduled_cutoff_time_this_run = list_predict_schedule[i][1]
-        required_time_this_run = performance_data_csv.get_value(instance, solver)
+        performance = sum(performance_list)
+    else:
+        # Minimum or maximum of predicted solvers (Aggregation function)
+        performance_list = []
+        for prediction in list_predict_schedule:
+            solver_performance = performance_data_csv.get_value(instance, prediction[0])
+            performance_list.append(float(solver_performance))
 
-        if required_time_this_run <= scheduled_cutoff_time_this_run:
-            used_time_for_this_instance = (
-                used_time_for_this_instance + required_time_this_run)
-            if used_time_for_this_instance > cutoff_time:
-                flag_successfully_solving = False
-            else:
+        if minimise:
+            performance = min(performance_list)
+            if performance <= capvalue:
                 flag_successfully_solving = True
-            break
         else:
-            used_time_for_this_instance = (
-                used_time_for_this_instance + scheduled_cutoff_time_this_run)
-            continue
+            performance = max(performance_list)
+            if performance >= capvalue:
+                flag_successfully_solving = True
 
-    return used_time_for_this_instance, flag_successfully_solving
+    return performance, flag_successfully_solving
 
 
 def compute_actual_selector_marginal_contribution(
+        aggregation_function: Callable[[list[float]], float],
+        capvalue_list: list[float],
+        minimise: bool,
         performance_data_csv_path: str = sgh.performance_data_csv_path,
         feature_data_csv_path: str = sgh.feature_data_csv_path,
         flag_recompute: bool = False) -> list[tuple[str, float]]:
     """Compute the marginal contributions of solvers in the selector.
 
     Args:
+      aggregation_function: Function to aggregate score values
+      capvalue_list: List of cap values
+      minimise: Flag indicating if scores should be minimised
       performance_data_csv: SparklePerformanceDataCSV object that holds the
         performance data.
       feature_data_csv_path: Path to the CSV file with the feature data.
@@ -372,17 +343,13 @@ def compute_actual_selector_marginal_contribution(
 
         return rank_list
 
-    cutoff_time_str = str(sgh.settings.get_general_target_cutoff_time())
-    print(f"In this calculation, cutoff time for each run is {cutoff_time_str} seconds")
+    print("In this calculation, cutoff time for each run is "
+          f"{sgh.settings.get_general_target_cutoff_time()} seconds")
 
     rank_list = []
 
     # Get values from CSV while all solvers and instances are included
     performance_data_csv = spdcsv.SparklePerformanceDataCSV(performance_data_csv_path)
-    num_instances = performance_data_csv.get_row_size()
-    num_solvers = performance_data_csv.get_column_size()
-    performance_measure = sgh.settings.get_general_performance_measure()
-    capvalue_list = get_capvalue_list(performance_data_csv, performance_measure)
 
     if not Path("Tmp/").exists():
         Path("Tmp/").mkdir()
@@ -395,20 +362,16 @@ def compute_actual_selector_marginal_contribution(
                                               feature_data_csv_path)
 
     if not Path(actual_portfolio_selector_path).exists():
-        print(f"****** WARNING: {actual_portfolio_selector_path} does not exist! ******")
-        print("****** WARNING: AutoFolio constructing the actual portfolio selector with"
+        print(f"****** ERROR: {actual_portfolio_selector_path} does not exist! ******")
+        print("****** ERROR: AutoFolio constructing the actual portfolio selector with"
               " all solvers failed! ******")
-        print("****** WARNING: Using virtual best performance instead of actual "
+        print("****** Use virtual best performance instead of actual "
               "performance for this portfolio selector! ******")
-        virtual_best_performance = (
-            performance_data_csv.calc_virtual_best_performance_of_portfolio(
-                num_instances, num_solvers, capvalue_list))
-        actual_selector_performance = virtual_best_performance
-    else:
-        actual_selector_performance = (
-            compute_actual_selector_performance(
-                actual_portfolio_selector_path, performance_data_csv_path,
-                feature_data_csv_path, num_instances, num_solvers, capvalue_list))
+        sys.exit(-1)
+
+    actual_selector_performance = compute_actual_selector_performance(
+        actual_portfolio_selector_path, performance_data_csv_path,
+        feature_data_csv_path, minimise, aggregation_function, capvalue_list)
 
     print("Actual performance for portfolio selector with all solvers is "
           f"{str(actual_selector_performance)}")
@@ -419,8 +382,8 @@ def compute_actual_selector_marginal_contribution(
         solver_name = sfh.get_last_level_directory_name(solver)
         print("Computing actual performance for portfolio selector excluding solver "
               f"{solver_name} ...")
-        tmp_performance_data_csv = (
-            spdcsv.SparklePerformanceDataCSV(performance_data_csv_path))
+        tmp_performance_data_csv = \
+            spdcsv.SparklePerformanceDataCSV(performance_data_csv_path)
         tmp_performance_data_csv.delete_column(solver)
         tmp_performance_data_csv_file = (
             f"tmp_performance_data_csv_without_{solver_name}_"
@@ -445,31 +408,37 @@ def compute_actual_selector_marginal_contribution(
             print("****** WARNING: No solver exists ! ******")
 
         if not Path(tmp_actual_portfolio_selector_path).exists():
-            print(f"****** WARNING: {tmp_actual_portfolio_selector_path} does not exist!"
+            print(f"****** ERROR: {tmp_actual_portfolio_selector_path} does not exist!"
                   " ******")
-            print("****** WARNING: AutoFolio constructing the actual portfolio selector "
+            print("****** ERROR: AutoFolio constructing the actual portfolio selector "
                   f"excluding solver {solver_name} failed! ******")
-            print("****** WARNING: Using virtual best performance instead of actual "
-                  "performance for this portfolio selector! ******")
-            tmp_virtual_best_performance = (
-                tmp_performance_data_csv.calc_virtual_best_performance_of_portfolio(
-                    num_instances, num_solvers, capvalue_list))
-            tmp_actual_selector_performance = tmp_virtual_best_performance
-        else:
-            tmp_actual_selector_performance = (
-                compute_actual_selector_performance(
-                    tmp_actual_portfolio_selector_path, tmp_performance_data_csv_path,
-                    feature_data_csv_path, num_instances, num_solvers, capvalue_list))
+            sys.exit(-1)
 
-        print(f"Actual performance for portfolio selector excluding solver {solver_name}"
-              f" is {str(tmp_actual_selector_performance)}")
+        tmp_asp = compute_actual_selector_performance(
+            tmp_actual_portfolio_selector_path, tmp_performance_data_csv_path,
+            feature_data_csv_path, minimise, aggregation_function, capvalue_list)
+
+        print(f"Actual performance for portfolio selector excluding solver "
+              f"{solver_name} is {str(tmp_asp)}")
         os.system("rm -f " + tmp_performance_data_csv_path)
         sl.add_output(tmp_performance_data_csv_path,
                       "[removed] Temporary performance data")
         print("Computing done!")
 
-        marginal_contribution = (
-            actual_selector_performance - tmp_actual_selector_performance)
+        # 1. If the performance remains equal, this solver has no contribution
+        # 2. If there is a performance decay without this solver, it has a contribution
+        # 3. If there is a performance improvement, we have a bad selector
+        if tmp_asp == actual_selector_performance:
+            marginal_contribution = 0.0
+        elif minimise and tmp_asp > actual_selector_performance or not minimise and\
+                tmp_asp < actual_selector_performance:
+            marginal_contribution = tmp_asp / actual_selector_performance
+        else:
+            print("****** WARNING DUBIOUS SELECTOR/SOLVER:"
+                  f" The omission of solver {solver_name} yields an improvement."
+                  "The selector improves better without this solver. It may be usefull"
+                  " to construct a portfolio without this solver.")
+            marginal_contribution = 0.0
 
         solver_tuple = (solver, marginal_contribution)
         rank_list.append(solver_tuple)
@@ -482,80 +451,28 @@ def compute_actual_selector_marginal_contribution(
     # Write actual selector contributions to file
     write_marginal_contribution_csv(actual_margi_cont_path, rank_list)
 
-#    os.system(r'rm -f ' + actual_portfolio_selector_path)
     return rank_list
 
 
-def print_rank_list(rank_list: list, mode: int) -> None:
+def print_rank_list(rank_list: list, mode: str) -> None:
     """Print the solvers ranked by marginal contribution.
 
     Args:
       rank_list: A list of 2-tuples as returned by function
         compute_actual_selector_marginal_contribution of the form
         (solver name, marginal contribution).
-      mode: This integer parameter determines the reference for the solver ranking.
-         Available options are 1 for 'perfect selector' and 2 for
-         'actual selector'.
+      mode: The marginal contribution mode used to calculate the rank.
+            Either Actual or Virtual.
     """
-    reference_selector = ""
-    if mode == 1:
-        reference_selector = "perfect selector"
-    elif mode == 2:
-        reference_selector = "actual selector"
-
     print("******")
     print("Solver ranking list via marginal contribution (Margi_Contr) with regards to "
-          f"{reference_selector}")
+          f"{mode}")
     for i in range(0, len(rank_list)):
         solver = rank_list[i][0]
         marginal_contribution = rank_list[i][1]
         print(f"#{str(i+1)}: {sfh.get_last_level_directory_name(solver)}\t Margi_Contr: "
               f"{str(marginal_contribution)}")
     print("******")
-
-
-def compute_perfect(flag_recompute: bool = False) -> list[tuple[str, float]]:
-    """Compute the marginal contribution for the perfect portfolio selector.
-
-    Args:
-        flag_recompute: Flag indicating whether marginal contributions
-            should be recalculated.
-
-    Returns:
-        rank_list: A list of 2-tuples of the form (solver name, marginal contribution).
-    """
-    print(
-        "Start computing each solver's marginal contribution to perfect selector ..."
-    )
-    rank_list = compute_perfect_selector_marginal_contribution(
-        flag_recompute=flag_recompute
-    )
-    print_rank_list(rank_list, 1)
-    print("Marginal contribution (perfect selector) computing done!")
-
-    return rank_list
-
-
-def compute_actual(flag_recompute: bool = False) -> list[tuple[str, float]]:
-    """Compute the marginal contribution for the actual portfolio selector.
-
-    Args:
-        flag_recompute: Flag indicating whether marginal contributions
-            should be recalculated.
-
-    Returns:
-        rank_list: A list of 2-tuples of the form (solver name, marginal contribution).
-    """
-    print(
-        "Start computing each solver's marginal contribution to actual selector ..."
-    )
-    rank_list = compute_actual_selector_marginal_contribution(
-        flag_recompute=flag_recompute
-    )
-    print_rank_list(rank_list, 2)
-    print("Marginal contribution (actual selector) computing done!")
-
-    return rank_list
 
 
 def compute_marginal_contribution(
@@ -571,11 +488,49 @@ def compute_marginal_contribution(
         flag_recompute: Flag indicating whether marginal contributions
             should be recalculated.
     """
-    if flag_compute_perfect:
-        compute_perfect(flag_recompute)
-    elif flag_compute_actual:
-        compute_actual(flag_recompute)
+    performance_data_csv = (
+        spdcsv.SparklePerformanceDataCSV(sgh.performance_data_csv_path))
+    performance_measure = sgh.settings.get_general_performance_measure()
+    aggregation_function = sgh.settings.get_general_metric_aggregation_function()
+    if performance_measure == PerformanceMeasure.QUALITY_ABSOLUTE_MAXIMISATION:
+        capvalue_list = sgh.settings.get_general_cap_value()
+        minimise = False
+    elif performance_measure == PerformanceMeasure.QUALITY_ABSOLUTE_MINIMISATION:
+        capvalue = sgh.settings.get_general_cap_value()
+        minimise = True
     else:
+        # assume runtime optimization
+        capvalue = sgh.settings.get_general_target_cutoff_time()
+        minimise = True
+
+    num_of_instances = performance_data_csv.get_number_of_instances()
+    if capvalue is list or capvalue is None:
+        capvalue_list = capvalue
+    else:
+        capvalue_list = [capvalue for _ in range(num_of_instances)]
+
+    if not (flag_compute_perfect | flag_compute_actual):
         print("ERROR: compute_marginal_contribution called without a flag set to"
               " True, stopping execution")
-        sys.exit()
+        sys.exit(-1)
+    if flag_compute_perfect:
+        print("Start computing each solver's marginal contribution "
+              "to perfect selector ...")
+        rank_list = compute_perfect_selector_marginal_contribution(
+            aggregation_function,
+            capvalue_list,
+            minimise,
+            flag_recompute=flag_recompute
+        )
+        print_rank_list(rank_list, "perfect selector")
+        print("Marginal contribution (perfect selector) computing done!")
+    if flag_compute_actual:
+        print("Start computing each solver's marginal contribution "
+              "to actual selector ...")
+        rank_list = compute_actual_selector_marginal_contribution(
+            aggregation_function,
+            capvalue_list, minimise,
+            flag_recompute=flag_recompute
+        )
+        print_rank_list(rank_list, "actual selector")
+        print("Marginal contribution (actual selector) computing done!")