+++ Commands/sparkle_help/sparkle_construct_portfolio_selector_help.py
@@ -170,10 +170,11 @@ def construct_sparkle_portfolio_selector(selector_path: Path,
 
     cutoff_time_str = str(cutoff_time)
     python_executable = sgh.python_executable
-    performance_measure = sgh.settings.get_general_performance_measure()
-    if performance_measure == PerformanceMeasure.RUNTIME:
+    perf_measure = sgh.settings.get_general_performance_measure()
+    if perf_measure == PerformanceMeasure.RUNTIME:
         objective_function = "--objective runtime"
-    elif performance_measure == PerformanceMeasure.QUALITY_ABSOLUTE:
+    elif perf_measure == PerformanceMeasure.QUALITY_ABSOLUTE_MAXIMISATION or\
+            perf_measure == PerformanceMeasure.QUALITY_ABSOLUTE_MINIMISATION:
         objective_function = "--objective solution_quality"
     else:
         print("ERROR: Unknown performance measure in "