+++ tests/test_sparkle_marginal_contribution_help.py
@@ -7,8 +7,6 @@ import platform
 
 from Commands.sparkle_help import sparkle_compute_marginal_contribution_help as scmch
 from Commands.sparkle_help.sparkle_feature_data_csv_help import SparkleFeatureDataCSV
-from Commands.sparkle_help import sparkle_performance_data_csv_help as spdcsv
-from Commands.sparkle_help.sparkle_settings import PerformanceMeasure
 from Commands.sparkle_help import sparkle_global_help as sgh
 from Commands.sparkle_help.sparkle_settings import Settings
 
@@ -53,47 +51,36 @@ class TestMarginalContribution(TestCase):
         assert result == output
         pth.unlink()
 
-    def test_get_cap_value_list(self: TestCase) -> None:
-        """Test for method get_cap_value_list."""
-        csv_obj = None
-        output = scmch.get_capvalue_list(csv_obj, PerformanceMeasure.RUNTIME)
-        assert output is None
-
-        csv_data = ",Solvers/MiniSAT,Solvers/CSCCSat\n"\
-                   "Instances/PTN/Ptn-7824-b03.cnf,3000.0,3000.0\n"\
-                   "Instances/PTN/Ptn-7824-b15.cnf,3000.0,28.1747\n"\
-                   "Instances/PTN/Ptn-7824-b05.cnf,3000.0,3000.0\n"\
-                   "Instances/PTN/Ptn-7824-b13.cnf,3000.0,9.98625\n"\
-                   "Instances/PTN/Ptn-7824-b21.cnf,117.589,0.107158\n"\
-                   "Instances/PTN/Ptn-7824-b19.cnf,3000.0,183.437\n"\
-                   "Instances/PTN/Ptn-7824-b17.cnf,3000.0,0.537186\n"\
-                   "Instances/PTN/bce7824.cnf,3000.0,3000.0\n"\
-                   "Instances/PTN/Ptn-7824-b01.cnf,3000.0,3000.0\n"\
-                   "Instances/PTN/Ptn-7824-b11.cnf,3000.0,3000.0\n"\
-                   "Instances/PTN/Ptn-7824-b09.cnf,3000.0,196.792\n"\
-                   "Instances/PTN/Ptn-7824-b07.cnf,3000.0,3000.0\n"
-        csv_path = Path("Test_Data/test_sparkle_performance_data.csv")
-        Path.write_text(csv_path, csv_data)
-        csv_obj = spdcsv.SparklePerformanceDataCSV(csv_path)
-        assert csv_obj.get_column_size() == 2
-        result = [3000.0, 3000.0, 3000.0, 3000.0, 117.589, 3000.0,
-                  3000.0, 3000.0, 3000.0, 3000.0, 3000.0, 3000.0]
-        output = scmch.get_capvalue_list(csv_obj, PerformanceMeasure.QUALITY_ABSOLUTE)
-
-        assert output == result
-        csv_path.unlink()
-
     def test_compute_perfect_selector_marginal_contribution(self: TestCase) -> None:
         """Test for method compute_perfect_selector_marginal_contribution."""
         pth = Path("Commands/test/test_files/Performance_Data/"
                    "test_construct_sparkle_portfolio_selector.csv")
 
-        # Strange behaviour, keeps changing its output 'randomly' between these two
-        resulta = [("Solvers/CSCCSat", 2.068482775510204), ("Solvers/MiniSAT", 0.0)]
-        resultb = [("Solvers/CSCCSat", 4.139621586398334), ("Solvers/MiniSAT", 0.0)]
+        # Settings have no impact yet on the unit test, this needs to be reconfigured
+        tmp_settings = ""
+        source_settings = Path("tests/test_files/Settings/mc-settings.ini")
+        with Path("Settings/sparkle_settings.ini").open("r+") as file_target:
+            tmp_settings = file_target.read()
+            file_target.truncate()
+            with source_settings.open("r") as file_source:
+                source_settings = file_source.read()
+                file_target.write(source_settings)
 
-        output = scmch.compute_perfect_selector_marginal_contribution(pth, True)
-        assert output == resulta or output == resultb
+        result = [("Solvers/CSCCSat", 1.7980089765503102), ("Solvers/MiniSAT", 0.0)]
+
+        output = scmch.compute_perfect_selector_marginal_contribution(
+            aggregation_function=sum,
+            capvalue_list=None,
+            minimise=True,
+            performance_data_csv_path=pth,
+            flag_recompute=True
+        )
+
+        with Path("Settings/sparkle_settings.ini").open("w") as f:
+            f.truncate()
+            f.write(tmp_settings)
+
+        self.assertListEqual(output, result)
 
     def test_get_list_predict_schedule(self: TestCase) -> None:
         """Test for method get_list_predict_schedule."""
@@ -129,18 +116,15 @@ class TestMarginalContribution(TestCase):
                     "test_construct_sparkle_portfolio_selector.csv"
         feature_csv_path = "Commands/test/test_files/Feature_Data/"\
                            "test_construct_sparkle_portfolio_selector.csv"
-        num_instances = 12
-        num_solvers = 2
-        capvalue_list = None
 
-        result = 2.068482775510204
+        result = 30.644344
         sgh.settings = Settings("Commands/test/test_files/sparkle_settings.ini")
         output = scmch.compute_actual_selector_performance(pth,
                                                            perf_path,
                                                            feature_csv_path,
-                                                           num_instances,
-                                                           num_solvers,
-                                                           capvalue_list)
+                                                           True,
+                                                           sum,
+                                                           None)
 
         assert output == result
 
@@ -148,28 +132,6 @@ class TestMarginalContribution(TestCase):
         """Test for method compute_actual_performance_for_instance."""
         # TODO: This method is currently not touched by the .sh test. Think of a test.
         pass
-
-    def test_compute_actual_used_time_for_instance(self: TestCase) -> None:
-        """Test for method compute_actual_used_time_for_instance."""
-        # Test does not work on Mac
-        # Does not work yet on Bitbucket due to issues with get_list..._from_file()
-        if platform.system() != "Linux" or is_bitbucket():
-            return
-        pth = "tests/data/sparkle_portfolio_selector__@@SPARKLE@@__"
-        instance = "Instances/PTN/Ptn-7824-b03.cnf"
-        perf_path = "Commands/test/test_files/Performance_Data/"\
-                    "test_construct_sparkle_portfolio_selector.csv"
-        feature_csv_path = "Commands/test/test_files/Feature_Data/"\
-                           "test_construct_sparkle_portfolio_selector.csv"
-        perfcsv = SparkleFeatureDataCSV(perf_path)
-
-        result = (3.0, False)
-
-        output = scmch.compute_actual_used_time_for_instance(pth,
-                                                             instance,
-                                                             feature_csv_path,
-                                                             perfcsv)
-        assert output == result
 
     def test_compute_actual_selector_marginal_contribution(self: TestCase) -> None:
         """Test for method compute_actual_selector_marginal_contribution."""
@@ -180,14 +142,18 @@ class TestMarginalContribution(TestCase):
                     "test_construct_sparkle_portfolio_selector.csv"
         feature_csv_path = "Commands/test/test_files/Feature_Data/"\
                            "test_construct_sparkle_portfolio_selector.csv"
-        flag_recompute = True
 
-        result = [("Solvers/CSCCSat", 4.139621586398334), ("Solvers/MiniSAT", 0.0)]
+        result = [("Solvers/CSCCSat", 1.3895076764357648), ("Solvers/MiniSAT", 0.0)]
 
-        output = scmch.compute_actual_selector_marginal_contribution(perf_path,
-                                                                     feature_csv_path,
-                                                                     flag_recompute)
-        assert output == result
+        output = scmch.compute_actual_selector_marginal_contribution(
+            aggregation_function=sum,
+            capvalue_list=None,
+            minimise=True,
+            performance_data_csv_path=perf_path,
+            feature_data_csv_path=feature_csv_path,
+            flag_recompute=True
+        )
+        self.assertEqual(output, result)
 
     def test_print_rank_list(self: TestCase) -> None:
         """Test for method print_rank_list. Could be irrelevant."""